backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
anthropic>=0.49
python-dotenv==1.0.1
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.1
httpx>=0.27,<0.28
apscheduler>=3.10
beautifulsoup4>=4.12
